requirements.txt
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
dacite==1.5.1
dj-database-url==0.5.0
Django==2.2.5
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==2.0.0
mysqlclient==1.4.6
pytz==2020.1
requests==2.22.0
responses==0.12.0
sentry-sdk==0.19.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
zipp==3.4.0
